+++ src/ray.rs
@@ -13,7 +13,7 @@ impl Ray {
         }
     }
 
-    pub fn at(&self, t: f64) -> Vec3 {
+    pub fn at(&self, t: f32) -> Vec3 {
         self.origin + self.direction * t
     }
 }
